requirements.txt
# Core dependencies
playwright==1.48.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pyyaml==6.0.2

# Video processing
moviepy==1.0.3
opencv-python==4.10.0.84

# Web UI
gradio==4.44.0

# Utilities
requests==2.32.3
aiohttp==3.10.10
tqdm==4.66.5
